+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
   - Add a new trigger action: "Execute Lua"
   - Fix an issue in melee unit initialization (introduced in 0.2.0)
+  - Add a command line argument `-keep-app-active` that keeps the app active in the background
 
 ## 0.2.0 - Busy Barracks
   - Reimplement ~10% of the functions in StarCraft 1.16.1